requirements.txt
Django==1.5.2
Jinja2==2.7
Markdown==2.3.1
MarkupSafe==0.18
Pillow==2.1.0
South==0.8.1
django-filter==0.6
django-widget-tweaks==1.3
djangorestframework==2.3.5
honcho==0.4.2
ipdb==0.7
ipython==0.13.2
py-bcrypt==0.3
requests==1.2.3
wsgiref==0.1.2
